+++ src/main/java/facade/HomeworkTwoFacade.java
@@ -31,7 +31,7 @@ public class HomeworkTwoFacade {
      * @return a new level determined by the parameters
      * @see Level
      */
-    public Level newLevelWithBricksFull(String name, int numberOfBricks, double probOfGlass, double probOfMetal, int seed) {
+    public Level newLevelWithBricksFull(String name, int numberOfBricks, double probOfGlass, double probOfMetal, long seed) {
         return game.newLevelWithBricksFull(name, numberOfBricks, probOfGlass, probOfMetal, seed);
     }
 
@@ -45,7 +45,7 @@ public class HomeworkTwoFacade {
      * @return a new level determined by the parameters
      * @see Level
      */
-    public Level newLevelWithBricksNoMetal(String name, int numberOfBricks, double probOfGlass, int seed) {
+    public Level newLevelWithBricksNoMetal(String name, int numberOfBricks, double probOfGlass, long seed) {
         return game.newLevelWithBricksNoMetal(name, numberOfBricks, probOfGlass, seed);
     }
